requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
deprecation==2.1.0
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.4.0
django-deep-serializer==0.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
filelock==3.15.4
fonttools==4.53.0
frozenlist==1.4.1
gotrue==2.9.2
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.7
kiwisolver==1.4.5
lxml==5.2.2
Markdown==3.6
matplotlib==3.9.0
multidict==6.1.0
numpy==1.26.4
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
postgrest==0.17.1
propcache==0.2.0
psutil==5.9.8
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
realtime==2.0.5
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
storage3==0.8.1
StrEnum==0.4.15
supabase==2.9.0
supafunc==0.6.1
tabulate==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.3
virtualenvwrapper-win==1.2.7
websockets==13.1
XlsxWriter==3.2.0
yarl==1.14.0
